+++ vm/src/vm/runners/builtin_runner/hash.rs
@@ -1,3 +1,4 @@
+use crate::air_private_input::{PrivateInput, PrivateInputPair};
 use crate::stdlib::{cell::RefCell, prelude::*};
 use crate::types::errors::math_errors::MathError;
 use crate::types::instance_definitions::pedersen_instance_def::{
@@ -186,6 +187,30 @@ impl HashBuiltinRunner {
             }
         }
         BuiltinAdditionalData::Hash(verified_addresses)
+    }
+
+    pub fn air_private_input(&self, memory: &Memory) -> Vec<PrivateInput> {
+        let mut private_inputs = vec![];
+        if let Some(segment) = memory.data.get(self.base) {
+            let segment_len = segment.len();
+            for (index, off) in (0..segment_len)
+                .step_by(CELLS_PER_HASH as usize)
+                .enumerate()
+            {
+                // Add the input cells of each hash instance to the private inputs
+                if let (Ok(x), Ok(y)) = (
+                    memory.get_integer((self.base as isize, off).into()),
+                    memory.get_integer((self.base as isize, off + 1).into()),
+                ) {
+                    private_inputs.push(PrivateInput::Pair(PrivateInputPair {
+                        index,
+                        x: *x,
+                        y: *y,
+                    }))
+                }
+            }
+        }
+        private_inputs
     }
 }
 
@@ -546,5 +571,44 @@ mod tests {
             builtin.get_additional_data(),
             BuiltinAdditionalData::Hash(verified_addresses)
         )
+    }
+
+    #[test]
+    #[cfg_attr(target_arch = "wasm32", wasm_bindgen_test)]
+    fn get_air_private_input() {
+        let builtin: BuiltinRunner = HashBuiltinRunner::new(None, true).into();
+
+        let memory = memory![
+            ((0, 0), 0),
+            ((0, 1), 1),
+            ((0, 2), 2),
+            ((0, 3), 3),
+            ((0, 4), 4),
+            ((0, 5), 5),
+            ((0, 6), 6),
+            ((0, 7), 7),
+            ((0, 8), 8),
+            ((0, 9), 9)
+        ];
+        assert_eq!(
+            builtin.air_private_input(&memory),
+            (vec![
+                PrivateInput::Pair(PrivateInputPair {
+                    index: 0,
+                    x: 0.into(),
+                    y: 1.into()
+                }),
+                PrivateInput::Pair(PrivateInputPair {
+                    index: 1,
+                    x: 3.into(),
+                    y: 4.into()
+                }),
+                PrivateInput::Pair(PrivateInputPair {
+                    index: 2,
+                    x: 6.into(),
+                    y: 7.into()
+                }),
+            ]),
+        );
     }
 }