+++ vm/src/vm/vm_core.rs
@@ -78,8 +78,8 @@ impl DeducedOperands {
 
 pub struct VirtualMachine {
     pub(crate) run_context: RunContext,
-    pub(crate) builtin_runners: Vec<BuiltinRunner>,
-    pub(crate) segments: MemorySegmentManager,
+    pub builtin_runners: Vec<BuiltinRunner>,
+    pub segments: MemorySegmentManager,
     pub(crate) trace: Option<Vec<TraceEntry>>,
     pub(crate) current_step: usize,
     pub(crate) rc_limits: Option<(isize, isize)>,
@@ -1076,6 +1076,23 @@ impl VirtualMachine {
                 ))
             })
             .collect()
+    }
+
+    #[doc(hidden)]
+    pub fn builtins_final_stack_from_stack_pointer_dict(
+        &mut self,
+        builtin_name_to_stack_pointer: &HashMap<&'static str, Relocatable>,
+    ) -> Result<(), RunnerError> {
+        for builtin in self.builtin_runners.iter_mut() {
+            builtin.final_stack(
+                &self.segments,
+                builtin_name_to_stack_pointer
+                    .get(builtin.name())
+                    .cloned()
+                    .unwrap_or_default(),
+            )?;
+        }
+        Ok(())
     }
 }